requirements.txt
numpy==1.20.3
pandas==1.3.4
Pillow==9.2.0
plotly==5.10.0
plotly_express==0.4.1
streamlit==1.12.0
